refac: apply single resposability in auth module
