+++ src/auth/controllers/sign-up.controller.ts
@@ -1,0 +1,17 @@
+import {
+    Body,
+    Controller,
+    Post,
+  } from '@nestjs/common';
+  import { SignUpService } from '../services/sign-up.service';
+  import { CreateUserDTO } from '../dtos';
+  
+  @Controller('auth')
+  export class SignUpController {
+    constructor(private signUpService: SignUpService) {}
+  
+    @Post('signup')
+    signUp(@Body() dto: CreateUserDTO) {
+      return this.signUpService.singUp(dto);
+    }
+  }
